requirements.txt
azure-ai-ml
azureml-core